third_party/WebKit/LayoutTests/http/tests/inspector/styles/selector-line-deprecated-expected.txt

CONSOLE MESSAGE: "/*@ sourceMappingURL=" source mapping URL declaration is deprecated, "/*# sourceMappingURL=" declaration should be used instead.
Tests that selector line is computed correctly regardless of its start column. Bug 110732.

Text
[expanded] 
element.style  { ()

======== Matched CSS Rules ========
[expanded] 
#container #inspected  { (selector-line.scss:4 -> selector-line.scss:4:5)
color: green;

[expanded] 
#inspected  { (selector-line-deprecated.html:4 -> selector-line-deprecated.html:4:1)
/-- overloaded --/ color: green;

[expanded] 
div  { (user agent stylesheet)
display: block;
